+++ pg_documentdb/include/metadata/metadata_cache.h
@@ -39,7 +39,7 @@ extern PGDLLIMPORT char *ApiAdminRoleV2;
 extern PGDLLIMPORT char *ApiBgWorkerRole;
 extern PGDLLIMPORT char *ApiReadOnlyRole;
 extern PGDLLEXPORT char *ApiReadWriteRole;
-extern PGDLLEXPORT char *ApiRootReplaceRole;
+extern PGDLLEXPORT char *ApiRootInternalRole;
 extern PGDLLIMPORT char *ApiRootRole;
 extern PGDLLEXPORT char *ApiUserAdminRole;